+++ controllers/userController.js
@@ -11,8 +11,9 @@ module.exports = {
   },
   async getSingleUser(req, res) {
     try {
-      const user = await User.findOne({ _id: req.params.userId })
-        .select('-__v');
+      const user = await User.findOne({ _id: req.params.userId }).select(
+        '-__v'
+      );
 
       if (!user) {
         return res.status(404).json({ message: 'No user with that ID' });
@@ -31,5 +32,27 @@ module.exports = {
     } catch (err) {
       res.status(500).json(err);
     }
+  },
+
+  // update a user
+  async updateUser(req, res) {
+    try {
+      const user = await User.findOneAndUpdate(
+        { _id: req.params.id },
+        { $addToSet: { thoughts: req.body } },
+        { runValidators: true, new: true }
+      );
+
+      if (!user) {
+        return res
+          .status(404)
+          .json({ message: 'No user found with that ID :(' });
+      }
+
+      res.json(user);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+   
   },
 };