+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 5. Updated Thoughts model to include Reactions
 6. Add User and Thought Routes
 7. Added Controllers - some not ready & so commented out
+8. Updated User Route & userController to add thought and update user